apps/api/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.4.1,<2.8
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
asyncpg>=0.29.0
sqlalchemy>=2.0.0
alembic>=1.12.0
PyJWT>=2.8.0          # ← ДОБАВИТЬ ЭТО
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aiohttp>=3.8.0
httpx>=0.25.0
pytoniq>=0.1.0
